pyproject.toml
[project]
name = "amcp-agent"
version = "0.5.0"
description = "amcp: Lego-style coding agent CLI. Core: rg grep/find and read; editing & web search via MCP."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "tao" }
]
license = { text = "MIT" }
keywords = ["mcp", "agent", "cli", "coding", "ai", "llm"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development",
]

dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "pydantic>=2.7",
  "fastmcp>=2.0.0,<3",
  "tomli>=2.0.0;python_version<'3.11'",
  "tomli-w>=1.0.0",
  "openai>=1.52.0",
  "pyyaml>=6.0",
  "prompt-toolkit>=3.0.0",
  "agent-client-protocol>=0.7.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.3.0",
  "mypy>=1.8.0",
  "anthropic>=0.40.0",
]

[project.scripts]
amcp = "amcp.cli:app"
amcp-acp = "amcp.acp_agent:main"

[project.urls]
Homepage = "https://github.com/tao12345666333/amcp"
Repository = "https://github.com/tao12345666333/amcp"
Issues = "https://github.com/tao12345666333/amcp/issues"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/amcp"]

[tool.hatch.build.targets.sdist]
include = ["src/amcp"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/amcp --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
